All methods put into different classes to clean things up. Checked human vs human and human vs computer and all worked
